[consensus] simplify pacemaker impl - sequential event processor cleanup part 3
