+++ consensus/src/chained_bft/liveness/mod.rs
@@ -1,7 +1,6 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-pub(crate) mod local_pacemaker;
 pub(crate) mod pacemaker;
 pub(crate) mod pacemaker_timeout_manager;
 pub(crate) mod proposal_generator;
@@ -9,6 +8,6 @@ pub(crate) mod proposer_election;
 pub(crate) mod rotating_proposer_election;
 
 #[cfg(test)]
-mod local_pacemaker_test;
+mod pacemaker_test;
 #[cfg(test)]
 mod rotating_proposer_test;
